refactor(admin-175): remove redundant admin actions and move Add Accessory
